+++ src/Common/Infrastructure/Infrastructure.csproj
@@ -16,7 +16,6 @@
 
   <ItemGroup>
     <Folder Include="Persistence\Migrations\" />
-    <Folder Include="Services\WeatherForeCast\" />
   </ItemGroup>
 
   <ItemGroup>
